Use dotenv for local development only
Configuring the app via a `.env` file is problematic because it affects
all environments, including test. The test environment should work
regardless of local configuration.

This commit more narrowly focuses the usage of dotenv by recommending a
`.env.development` file be used as opposed to `.env`. This limits the
effects to just the local development environment.

Also, there is no reason to load the dotenv gem in production.