Make SSL opt-out instead of opt-in
Before, SSL would be disabled in production by default, and you would
need to set RAILS_FORCE_SSL=true to turn in on.

Now, SSL is on by default, and you can turn if off with
RAILS_DISABLE_SSL=true.

This more closely aligns with the upcoming Rails 7.1, which will have
SSL enabled by default in the production config.